Integrate CAPIF wiht backoffice and update test to new deploy system

+++ services/TS29222_CAPIF_Publish_Service_API/published_apis/core/consumer_messager.py
@@ -19,9 +19,10 @@ class Subscriber():
     def listen(self):
         current_app.logger.info("Listening publish messages")
         for raw_message in self.p.listen():
+            current_app.logger.debug(f"Received message")
             if raw_message["type"] == "message" and raw_message["channel"].decode('utf-8') == "internal-messages":
                 message, *ids = raw_message["data"].decode('utf-8').split(":")
-                if message == "provider-removed":
+                if message == "provider-removed" and len(ids)==2:
                     self.security_ops.delete_intern_service(ids[1])
